클릭시 숫자 표시 하는 것 풀링 + 이미지 교체해주는 것

+++ Assets/Scripts/Base/User.cs
@@ -5,6 +5,6 @@ public class User
 {
     public string nickname;
     public long exp;
-    public long clickExp;
+    public int clickExp;
     public List<Place> placeList = new List<Place>();
 }
